+++ src/lib/panel-main/display/call-item-slim.svelte
@@ -10,7 +10,8 @@
 
 	export let time: string
 	export let call: CallItem
-	export let style: { default_text: string; icon: Component ; classes: string }
+	export let style: { default_text: string; icon: Component; classes: string }
+	export let volume: number
 </script>
 
 <div