+++ src/lib/panel-main/display/call.svelte
@@ -7,10 +7,13 @@
 		IconPhoneOutgoing,
 		IconPlugConnected
 	} from '@tabler/icons-svelte'
-	import { onDestroy, onMount } from 'svelte'
+	import { getContext, onDestroy, onMount } from 'svelte'
 	import { config } from '$lib/stores/config.persistent'
 	import CallItemDefault from './call-item-default.svelte'
 	import CallItemSlim from './call-item-slim.svelte'
+	import type Manager from '$lib/client-phone'
+
+	$: ({ dialpad_enabled, dialpad_extended, call_level_indicator_disabled } = $config)
 
 	export let index: number
 	export let call: CallItem
@@ -78,7 +81,7 @@
 		}, 1000)
 	})
 
-	$: height_call_compact = ($config.dialpad_enabled && $config.dialpad_extended ? 60 : 56) + 4
+	$: height_call_compact = (dialpad_enabled && dialpad_extended ? 60 : 56) + 4
 	$: not_slim = height_display >= height_call_compact * $calls.length + 4
 
 	const handleMouseUp = (e: MouseEvent) => e.button !== 1 || (call.selected = !call.selected)
@@ -128,6 +131,10 @@
 			}
 		}
 	}
+
+	let volume = 0
+	const phone = getContext<Manager>('phone')
+	phone.on('volume', ({ id, value }) => id !== call.id || (volume = value))
 </script>
 
 <button
@@ -141,11 +148,24 @@
 	on:keydown={handleKeydown}
 >
 	<div class="flex {not_slim ? 'hidden' : 'call-slim'}">
-		<CallItemSlim bind:call bind:time bind:style />
+		<CallItemSlim bind:call bind:time bind:style {volume} />
 	</div>
 	<div class="flex {not_slim ? '' : 'call-default'} ">
-		<CallItemDefault bind:call bind:time bind:style />
+		<CallItemDefault bind:call bind:time bind:style {volume} />
 	</div>
+	{#if !call_level_indicator_disabled}
+		<div
+			class="
+				h-0 relative bottom-0.5 flex justify-center transition duration-75
+				{call.selected ? '' : 'opacity-70'}
+				"
+		>
+			<div
+				style="opacity: {volume * 1.9}%; width: {volume}%;"
+				class="h-[2px] bg-black dark:bg-white rounded-full backdrop-"
+			/>
+		</div>
+	{/if}
 </button>
 
 <style>